common utility functions for flex and flow reorder
