+++ editor/src/components/canvas/canvas-strategies/flow-reorder-strategy.tsx
@@ -1,36 +1,18 @@
 import { MetadataUtils } from '../../../core/model/element-metadata-utils'
-import * as EP from '../../../core/shared/element-path'
-import { offsetPoint } from '../../../core/shared/math-utils'
-import { CSSCursor } from '../canvas-types'
-import { reorderElement } from '../commands/reorder-element-command'
-import { setCursorCommand } from '../commands/set-cursor-command'
-import { setElementsToRerenderCommand } from '../commands/set-elements-to-rerender-command'
-import { updateHighlightedViews } from '../commands/update-highlighted-views-command'
 import { ParentBounds } from '../controls/parent-bounds'
 import { ParentOutlines } from '../controls/parent-outlines'
 import {
   CanvasStrategy,
-  emptyStrategyApplicationResult,
   getTargetPathsFromInteractionTarget,
   InteractionCanvasState,
   strategyApplicationResult,
-  StrategyApplicationResult,
 } from './canvas-strategy-types'
-import { absolute } from '../../../utils/utils'
-import { InteractionSession, StrategyState } from './interaction-state'
+import { InteractionSession } from './interaction-state'
 import { ElementInstanceMetadataMap } from '../../../core/shared/element-template'
-import {
-  getFlowReorderIndex,
-  getNewDisplayTypeForIndex,
-  getOptionalDisplayPropCommands,
-  isValidFlowReorderTarget,
-} from './flow-reorder-helpers'
-import {
-  FlowReorderAreaIndicator,
-  FlowReorderDragOutline,
-} from '../controls/flow-reorder-indicators'
+import { getOptionalDisplayPropCommands, isValidFlowReorderTarget } from './flow-reorder-helpers'
+import { FlowReorderDragOutline } from '../controls/flow-reorder-indicators'
 import { AllElementProps } from '../../editor/store/editor-state'
-import { isReorderAllowed } from './reorder-utils'
+import { applyReorderCommon } from './reorder-utils'
 
 function isFlowReorderConversionApplicable(
   canvasState: InteractionCanvasState,
@@ -46,7 +28,7 @@ function isFlowReorderConversionApplicable(
     if (
       siblings.length > 1 &&
       MetadataUtils.isPositionedByFlow(elementMetadata) &&
-      isValidFlowReorderTarget(elementMetadata)
+      isValidFlowReorderTarget(target, metadata)
     ) {
       return siblings.some((sibling) => MetadataUtils.isPositionedByFlow(sibling))
     } else {
@@ -54,80 +36,6 @@ function isFlowReorderConversionApplicable(
     }
   } else {
     return false
-  }
-}
-
-function flowReorderApplyCommon(
-  canvasState: InteractionCanvasState,
-  interactionState: InteractionSession,
-  strategyState: StrategyState,
-): StrategyApplicationResult {
-  if (interactionState.interactionData.type !== 'DRAG') {
-    return emptyStrategyApplicationResult
-  }
-
-  if (interactionState.interactionData.drag != null) {
-    const selectedElements = getTargetPathsFromInteractionTarget(canvasState.interactionTarget)
-    const target = selectedElements[0] // TODO MULTISELECT??
-
-    const siblingsOfTarget = MetadataUtils.getSiblingsProjectContentsOrdered(
-      strategyState.startingMetadata,
-      target,
-    ).map((element) => element.elementPath)
-
-    if (!isReorderAllowed(siblingsOfTarget)) {
-      return strategyApplicationResult(
-        [setCursorCommand('mid-interaction', CSSCursor.NotPermitted)],
-        {},
-        'failure',
-      )
-    }
-
-    const rawPointOnCanvas = offsetPoint(
-      interactionState.interactionData.dragStart,
-      interactionState.interactionData.drag,
-    )
-
-    const unpatchedIndex = siblingsOfTarget.findIndex((sibling) => EP.pathsEqual(sibling, target))
-    const lastReorderIdx = strategyState.customStrategyState.lastReorderIdx ?? unpatchedIndex
-
-    const { newIndex, targetSiblingUnderMouse } = getFlowReorderIndex(
-      interactionState.latestMetadata,
-      strategyState.startingAllElementProps,
-      rawPointOnCanvas,
-      target,
-    )
-
-    const newIndexFound = newIndex > -1
-    const mouseStillOverPreviousTargetSibling = EP.pathsEqual(
-      targetSiblingUnderMouse,
-      strategyState.customStrategyState.previousReorderTargetSiblingUnderMouse,
-    )
-
-    const newResultOrLastIndex =
-      !mouseStillOverPreviousTargetSibling && newIndexFound ? newIndex : lastReorderIdx
-
-    const newDisplayType = getNewDisplayTypeForIndex(
-      strategyState.startingMetadata,
-      target,
-      siblingsOfTarget[newResultOrLastIndex],
-    )
-
-    return strategyApplicationResult(
-      [
-        reorderElement('always', target, absolute(newResultOrLastIndex)),
-        setElementsToRerenderCommand(siblingsOfTarget),
-        updateHighlightedViews('mid-interaction', []),
-        setCursorCommand('mid-interaction', CSSCursor.Move),
-        ...getOptionalDisplayPropCommands(target, newDisplayType),
-      ],
-      {
-        lastReorderIdx: newResultOrLastIndex,
-        previousReorderTargetSiblingUnderMouse: targetSiblingUnderMouse,
-      },
-    )
-  } else {
-    return strategyApplicationResult([setCursorCommand('mid-interaction', CSSCursor.Move)])
   }
 }
 
@@ -164,5 +72,23 @@ export const flowReorderStrategy: CanvasStrategy = {
       ? 3
       : 0
   },
-  apply: flowReorderApplyCommon,
+  apply: (canvasState, interactionState, strategyState) => {
+    const reorderResult = applyReorderCommon(
+      canvasState,
+      interactionState,
+      strategyState,
+      isValidFlowReorderTarget,
+    )
+
+    const commands = [
+      ...reorderResult.commands,
+      ...getOptionalDisplayPropCommands(
+        reorderResult.customStatePatch.lastReorderIdx,
+        canvasState.interactionTarget,
+        strategyState.startingMetadata,
+      ),
+    ]
+
+    return strategyApplicationResult(commands, reorderResult.customStatePatch, reorderResult.status)
+  },
 }
